requirements.txt
# ═══════════════════════════════════════════════════════════════
# Playwright Test Scaffold - Dependencies
# ═══════════════════════════════════════════════════════════════

# ─────────────────────────────────────────────────────────────────
# Core Testing (Python 3.9+)
# ─────────────────────────────────────────────────────────────────
playwright==1.49.0
pytest==7.4.4
pytest-playwright==0.7.1

# ─────────────────────────────────────────────────────────────────
# Test Execution
# ─────────────────────────────────────────────────────────────────
pytest-xdist==3.8.0
pytest-rerunfailures==16.0.1   # retry on failure

# ─────────────────────────────────────────────────────────────────
# Reporting
# ─────────────────────────────────────────────────────────────────
allure-pytest==2.15.0
pytest-html>=4.1.1

# ─────────────────────────────────────────────────────────────────
# CLI & Console
# ─────────────────────────────────────────────────────────────────
click>=8.1.0
rich>=13.0.0

# ─────────────────────────────────────────────────────────────────
# Configuration
# ─────────────────────────────────────────────────────────────────
pyyaml>=6.0
python-dotenv==1.0.0

pydantic>=2.9.0
requests>=2.31.0